+++ lib/flutter_form_builder.dart
@@ -1,6 +1,5 @@
 library flutter_form_builder;
 
-// export 'package:country_code_picker/country_code_picker.dart';
 export 'package:flutter_typeahead/flutter_typeahead.dart';
 export 'package:signature/signature.dart';
 export 'package:dropdown_search/dropdown_search.dart' hide ErrorBuilder;
@@ -14,10 +13,7 @@ export './src/fields/form_builder_color_picker.dart';
 export './src/fields/form_builder_date_range_picker.dart';
 export './src/fields/form_builder_date_time_picker.dart';
 export './src/fields/form_builder_dropdown.dart';
-export './src/fields/form_builder_file_picker.dart';
 export './src/fields/form_builder_filter_chips.dart';
-export './src/fields/form_builder_image_picker.dart';
-export './src/fields/form_builder_phone_field.dart';
 export './src/fields/form_builder_radio_group.dart';
 export './src/fields/form_builder_range_slider.dart';
 export './src/fields/form_builder_rating.dart';